+++ src/app/questionaire-tab/questionaire-tab-routing.module.ts
@@ -1,16 +1,27 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
+import { QuestionaireGuard } from '../shared/questionaire.guard';
 
 import { QuestionaireTabPage } from './questionaire-tab.page';
 
 const routes: Routes = [
   {
-    path: '',
-    component: QuestionaireTabPage
+    path: 'dev',
+    component: QuestionaireTabPage,
+    data: { questionaire: "DEVELOPER" }
+  },
+  {
+    path: 'gen',
+    component: QuestionaireTabPage,
+    data: { questionaire: "GENERAL" }
   },
   {
     path: 'results',
     loadChildren: () => import('./results/results.module').then(m => m.ResultsPageModule)
+  },
+  {
+    path: '',
+    canActivate: [QuestionaireGuard]
   }
 ];
 
@@ -18,4 +29,4 @@ const routes: Routes = [
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule],
 })
-export class QuestionaireTabPageRoutingModule {}
+export class QuestionaireTabPageRoutingModule { }